Add web-platform-tests/wpt-tc-checks to the checks/statuses workaround
WPT is moving towards using the checks API for TaskCluster. However we
also use a decision task and as per
https://github.com/mozilla/community-tc-config/pull/129 this requires us
to be added to the route workaround.

Note that wpt-tc-checks is a temporary clone of wpt being used as we
develop the TaskCluster Checks integration; ultimately this will be
replaced with the wpt repo itself.