Reduce minCapacity to 0 everywhere
This will have the side-effect of slightly longer run times as we wait
for workers to be provisioned, but looking at the costs for time spent
idle, it would be hard to justify *not* making this change.

I've also explicitly added the minCapacity in some places where it was
missing (e.g. projects/taskcluster) instead of relying on the default
behavior. This will make it clearer when we look back when the change
was made.

While not stricly necessary, I've also reduced the maxCapacity for many
worker pools. The volume of jobs just doesn't justify the previous
requested max values.