Simplified hook definition:
The extra section is not needed - the description is available from the indexed
task and the expiration is automatically taken from the task expiration.

The python3 and docker version checks are unneeded as the task does not use
either.

The two task commands can be combined into one.

A file artifact is preferred over a directory artifact as it is more explicit
and avoids accidentally publishing additional artifacts accidentally.