Create pool proj-git-cinnabar/windows as a copy of proj-git-cinnabar/win2012r2
The intention here is to later remove proj-git-cinnabar/win2012r2
but keep both while tasks are being transitioned from one pool to
the other.

Removing the Windows version from the worker pool name will make
it less work to migrate to a newer Windows release in future.
This matches the naming of proj-git-cinnabar/linux for example.